finance_bot_project/requirements.txt
python-telegram-bot==22.7
python-dotenv==1.1.1
